Move encoder to room
